Everything a worker: cleanup queue/work mechanics.
